Rename files from form --> step
